backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-dotenv==1.0.1
httpx==0.27.0
pydantic[email]==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
greenlet
fastapi-mail
websockets==12.0
openai==1.30.1
apscheduler==3.10.4
